requirements.txt
openai>=1.13.3
requests>=2.31.0
python-dotenv>=1.0.1
ollama